+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-- fix config bug
+- added LibNolij (should shrink the JAR a bit)
 - further improvements to overall system stability and other minor adjustments have been made to enhance the user experience